added local_access() and read_access() methods to return of square-bracket operator; fixed consistency of square-bracket operator return handle with copies of that handle; locally owned access handles (for collection output on single index) are still a work-in-progress

+++ src/include/darma/impl/index_range/range_1d.h
@@ -193,9 +193,9 @@ struct basic_integer_range_1d
 
 } // end namespace detail
 
-using index_range_1d = detail::basic_integer_range_1d<int>;
+using index_range_1d = detail::basic_integer_range_1d<int64_t>;
 
-using index_1d = detail::basic_integer_index_1d<int>;
+using index_1d = detail::basic_integer_index_1d<int64_t>;
 
 } // end namespace indexing
